research/nlp/tinybert/requirements.txt
easydict
scipy>=1.7.1
onnxruntime-gpu